requirements.txt
streamlit==1.30.0
langchain==0.2.0
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
openai>=0.27.0
sentence-transformers
PyMuPDF==1.22.5
faiss-cpu
python-dotenv
streamlit_chat
pypdfium2
